TheCascadeBot/requirements.txt
discord.py>=2.3.0
asyncpg>=0.27.0
redis>=4.5.0
python-dotenv>=1.0.0
psutil>=5.9.0
aiohttp>=3.8.0